Refactors Exec to handle fg and bg commands
I removed RunCommandBg entirely and replaced it with Exec because both
function essentailly do the same thing with a slight difference in the
handling of the parent process. I chose to pass cmd->bg as a parameter
instead of having exec look for it because the it makes Exec's
functionality more generic.