Adds support for IO Redirection outwards
This commit adds support to redirect stdout to a file. This is done by
adding an additional parameter to the commandT struct that includes
the io redirection status of the command. If redirection required
(parsed in runtime.c:RunCmdForm), it sets cmd->ioRedirection to
the file descriptor (0 for stdin and 1 for stdout) that needs to
replaced and cmd->ioRedirectionPath to the file that the existing
file descriptor will be placed with. If cmd->ioRedirection is
detected in the child process prior to the exec, it switches the
file descriptors inside of exec.